Scooter#near(lat: 30.269362, lon: -97.7408848, radius: 75, after: '2016-06-16-03:58:41') => [Scooter, Scooter, Scooter, ...]
Note that I tried to make use of Rails Active Record sanitizers, but they actually screwed up the syntax of the geospatial query so I had to toss it. I'm still forcing radius to be a whole number by stripping any non digit character, but this could/should be improved for more precise queries.